Change throttle implementation
